backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0

# AI 模型 SDK（根據需要選擇）
dashscope>=1.12.0  # Qwen
openai>=1.3.0      # GPT-4V
anthropic>=0.7.0   # Claude
